ch(refactor): revamp to mvp architechture
- add more interface
- restructure service package

[Finishes #96743548]
