+++ app/src/main/AndroidManifest.xml
@@ -12,21 +12,21 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".DetailActivity"
+        <activity android:name=".view.DetailActivity"
             android:label = "Detail Activity"
-            android:parentActivityName=".MainActivity"
+            android:parentActivityName=".view.MainActivity"
             >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=
-                    "com.example.adekunleoluwafemi.javadevelopers.MainActivity" />
+                    "com.example.adekunleoluwafemi.javadevelopers.view.MainActivity" />
         </activity>
     </application>
